requirements.txt
requests==2.22.0
dataclasses==0.6
freezegun==0.3.15
psycopg2==2.8.5
kafka-python==2.0.1
flake8==3.8.2
pylint==2.5.2